refactor: клиент vps
