+++ src/client.ts
@@ -1,71 +1,7 @@
-import {
-    Message,
-    Settings,
-    SystemMessage,
-    Device,
-    IDevice,
-    ISettings,
-    Text,
-    Voice,
-    LegacyDevice,
-    ILegacyDevice,
-    InitialSettings,
-    IInitialSettings,
-} from './proto';
-import {
-    EventsType,
-    SystemMessageDataType,
-    VpsVersion,
-    CreateClientDataType,
-    ClientLogger,
-    MessageNames,
-} from './typings';
+import { Message, IDevice, ISettings, ILegacyDevice, IInitialSettings } from './proto';
+import { EventsType, SystemMessageDataType, CreateClientDataType, ClientLogger, MessageNames } from './typings';
 import { createNanoEvents } from './nanoevents';
-
-export const appendHeader = (buffer: Uint8Array) => {
-    // Добавляем 4 байта в начало с длинной сообщения
-    const arrayBuffer = new ArrayBuffer(4);
-    const dataView = new DataView(arrayBuffer, 0);
-    dataView.setInt32(0, buffer.length, true);
-    const uint8Array = new Uint8Array(4 + buffer.length);
-    uint8Array.set(new Uint8Array(arrayBuffer));
-    uint8Array.set(buffer, 4);
-
-    return uint8Array;
-};
-
-const compileBasePayload = ({
-    userId,
-    token,
-    userChannel,
-    version,
-    messageName,
-    vpsToken,
-}: {
-    userId: string;
-    token: string;
-    userChannel: string;
-    version: VpsVersion;
-    messageName?: string;
-    vpsToken?: string;
-}) => {
-    if (version < 3) {
-        return {
-            userId,
-            token,
-            userChannel,
-            messageName,
-            vpsToken,
-            version,
-        };
-    }
-
-    return {
-        token,
-        messageName,
-        version,
-    };
-};
+import { createClientMethods } from './clientMethods';
 
 export const createClient = (clientParams: CreateClientDataType, logger?: ClientLogger) => {
     const {
@@ -82,7 +18,6 @@ export const createClient = (clientParams: CreateClientDataType, logger?: Client
         vpsToken,
         meta,
     } = clientParams;
-    const basePayload = compileBasePayload({ userId, token, messageName, vpsToken, userChannel, version });
 
     let status: 'connecting' | 'ready' | 'closed' = 'connecting';
 
@@ -90,8 +25,8 @@ export const createClient = (clientParams: CreateClientDataType, logger?: Client
 
     const { on, emit, once } = createNanoEvents<EventsType>();
 
-    const pendingMessages: Map<any, Array<Message>> = new Map();
-    const commitedMessages: Map<any, Array<Message>> = new Map();
+    const pendingMessages: Map<number | Long, Array<Message>> = new Map();
+    const commitedMessages: Map<number | Long, Array<Message>> = new Map();
 
     let currentSettings = { device, legacyDevice, settings, locale };
     let currentMessageId = Date.now();
@@ -119,67 +54,41 @@ export const createClient = (clientParams: CreateClientDataType, logger?: Client
         });
     };
 
-    const send = ({
-        payload,
-        messageId,
-        ...other
-    }: {
-        payload: (
-            | { settings: Settings }
-            | { device: Device }
-            | { systemMessage: SystemMessage }
-            | { text: Text }
-            | { voice: Voice }
-            | { legacyDevice: LegacyDevice }
-            | { initialSettings: InitialSettings }
-        ) & {
-            last: 1 | -1;
-            messageName?: string;
-            meta?: { [k: string]: string };
-        };
-        messageId: number;
-    }) => {
-        const message = Message.create({
-            messageName: '',
-            // vpsToken: '',
-            ...basePayload,
-            ...payload,
-            messageId,
-            ...other,
-        });
-
-        const buffer = Message.encode(message).finish();
-        const bufferWithHeader = appendHeader(buffer);
-
+    const sendMessage = (message: Message, buffer: Uint8Array) => {
         logger?.logOutcoming(message);
 
         emit('outcoming', message);
 
         if (status === 'ready') {
-            ws.send(bufferWithHeader);
+            ws.send(buffer);
         } else {
-            messageQueue.push(bufferWithHeader);
+            messageQueue.push(buffer);
         }
     };
 
-    const sendDevice = (data: IDevice, last = true, messageId = getMessageId()) => {
+    const {
+        send,
+        sendDevice: sendDeviceOriginal,
+        sendInitialSettings: sendInitialSettingsOriginal,
+        sendLegacyDevice: sendLegacyDeviceOriginal,
+        sendSettings: sendSettingsOriginal,
+        sendText,
+        sendSystemMessage,
+        sendVoice,
+        updateDefaults,
+        batch,
+    } = createClientMethods(
+        { userId, token, messageName, vpsToken, userChannel, version },
+        { getMessageId, sendMessage, waitForAnswerToUser },
+    );
+
+    const sendDevice = ((data: IDevice, ...args: never[]) => {
         currentSettings = { ...currentSettings, device: data };
 
-        return send({
-            payload: {
-                device: Device.create(data),
-                last: last ? 1 : -1,
-            },
-            messageId,
-        });
-    };
+        return sendDeviceOriginal(data, ...args);
+    }) as typeof sendDeviceOriginal;
 
-    const sendInitialSettings = (
-        data: IInitialSettings,
-        last = true,
-        messageId = getMessageId(),
-        params: { meta?: { [k: string]: string } } = {},
-    ) => {
+    const sendInitialSettings = ((data: IInitialSettings, ...args: never[]) => {
         if (data.device && data.settings) {
             currentSettings = {
                 ...currentSettings,
@@ -189,117 +98,20 @@ export const createClient = (clientParams: CreateClientDataType, logger?: Client
             };
         }
 
-        return send({
-            payload: {
-                initialSettings: InitialSettings.create(data),
-                last: last ? 1 : -1,
-                ...params,
-            },
-            messageId,
-        });
-    };
+        return sendInitialSettingsOriginal(data, ...args);
+    }) as typeof sendInitialSettingsOriginal;
 
-    const sendLegacyDevice = (data: ILegacyDevice, last = true, messageId = getMessageId()) => {
+    const sendLegacyDevice = ((data: ILegacyDevice, ...args: never[]) => {
         currentSettings = { ...currentSettings, legacyDevice: data };
 
-        return send({
-            payload: {
-                legacyDevice: LegacyDevice.create(data),
-                last: last ? 1 : -1,
-            },
-            messageId,
-        });
-    };
+        return sendLegacyDeviceOriginal(data, ...args);
+    }) as typeof sendLegacyDeviceOriginal;
 
-    const sendSettings = (data: ISettings, last = true, messageId = getMessageId()) => {
+    const sendSettings = ((data: ISettings, ...args: never[]) => {
         currentSettings = { ...currentSettings, settings: data };
 
-        return send({
-            payload: {
-                settings: Settings.create(data),
-                last: last ? 1 : -1,
-            },
-            messageId,
-        });
-    };
-
-    const sendText = (
-        data: string,
-        params: {
-            messageId?: number;
-            last?: 1 | -1;
-            messageName?: string;
-            vpsToken?: string;
-            userId?: string;
-            token?: string;
-            userChannel?: string;
-            version?: VpsVersion;
-            meta?: { [k: string]: string };
-        } = {},
-        type = '',
-        messageId = getMessageId(),
-    ) => {
-        const text = type ? { data, type } : { data };
-        send({
-            payload: {
-                text: Text.create(text),
-                last: params.last ?? 1,
-            },
-            messageId,
-            ...params,
-        });
-
-        return waitForAnswerToUser(messageId);
-    };
-
-    const sendSystemMessage = (
-        { data, messageName = '' }: { data: Record<string, any>; messageName?: string },
-        last = true,
-        messageId = getMessageId(),
-        params: {
-            meta?: { [k: string]: string };
-        } = {},
-    ) => {
-        send({
-            payload: {
-                systemMessage: SystemMessage.create({
-                    data: JSON.stringify(data),
-                }),
-                messageName,
-                last: last ? 1 : -1,
-                ...params,
-            },
-            messageId,
-        });
-
-        return waitForAnswerToUser(messageId);
-    };
-
-    const sendVoice = (
-        data: Uint8Array,
-        last = true,
-        messageId = getMessageId(),
-        messageName?: string,
-        params: {
-            meta?: { [k: string]: string };
-        } = {},
-    ) => {
-        return send({
-            payload: {
-                voice: Voice.create({
-                    data: new Uint8Array(data),
-                }),
-                messageName,
-                last: last ? 1 : -1,
-                ...params,
-            },
-            messageId,
-        });
-    };
-
-    const updateDefauls = (obj: Partial<typeof basePayload>) => {
-        Object.assign(basePayload, obj);
-    };
+        return sendSettingsOriginal(data, ...args);
+    }) as typeof sendSettingsOriginal;
 
     const destroy = () => {
         destroyed = true;
@@ -348,7 +160,9 @@ export const createClient = (clientParams: CreateClientDataType, logger?: Client
 
                 while (messageQueue.length > 0) {
                     const message = messageQueue.shift();
-                    ws.send(message!);
+                    if (message) {
+                        ws.send(message);
+                    }
                 }
             }
 
@@ -398,102 +212,6 @@ export const createClient = (clientParams: CreateClientDataType, logger?: Client
 
     startWebSocket();
 
-    type BatchableMethods = {
-        sendDevice: typeof sendDevice;
-        sendLegacyDevice: typeof sendLegacyDevice;
-        sendSettings: typeof sendSettings;
-        sendInitialSettings: typeof sendInitialSettings;
-        send: typeof send;
-        sendText: typeof sendText;
-        sendSystemMessage: (
-            data: { data: Record<string, any>; messageName?: string },
-            last: boolean,
-            params?: {
-                meta?: { [k: string]: string };
-            },
-        ) => void;
-        sendVoice: (
-            data: Uint8Array,
-            last: boolean,
-            messageName?: string,
-            params?: {
-                meta?: { [k: string]: string };
-            },
-        ) => void;
-        messageId: number;
-    };
-
-    const batch = <T>(cb: (methods: BatchableMethods) => T): T => {
-        const batchingMessageId = getMessageId();
-        let lastMessageSent = false;
-        const checkLastMessageStatus = (last?: boolean) => {
-            if (lastMessageSent) {
-                if (last) {
-                    throw new Error("Can't send two last items in batch");
-                } else {
-                    throw new Error("Can't send messages in batch after last message have been sent");
-                }
-            } else if (last) {
-                lastMessageSent = true;
-            }
-        };
-
-        const threeParamsMethods = Object.entries({
-            sendDevice,
-            sendSettings,
-            sendInitialSettings,
-            sendLegacyDevice,
-        }).reduce((acc, curr) => {
-            const key = curr[0] as 'sendDevice' | 'sendSettings' | 'sendInitialSettings' | 'sendLegacyDevice';
-            acc[key] = (...params: Parameters<typeof curr[1]>) => {
-                checkLastMessageStatus(params[1]);
-                return curr[1](params[0], params[1], batchingMessageId);
-            };
-            return acc;
-        }, {} as Pick<BatchableMethods, 'sendDevice' | 'sendSettings' | 'sendInitialSettings' | 'sendLegacyDevice'>);
-        const upgradedSend: typeof send = (params) => {
-            checkLastMessageStatus(params.payload.last === 1);
-            return send({ ...params, messageId: batchingMessageId });
-        };
-
-        const upgradedSendText: typeof sendText = (...[data, params, type]) => {
-            checkLastMessageStatus(params?.last === 1);
-            return sendText(data, params, type, batchingMessageId);
-        };
-
-        const upgradedSendSystemMessage: (
-            data: { data: Record<string, any>; messageName?: string },
-            last: boolean,
-            params?: {
-                meta?: { [k: string]: string };
-            },
-        ) => void = (data, last, params) => {
-            checkLastMessageStatus(last);
-            return sendSystemMessage(data, last, batchingMessageId, params);
-        };
-
-        const upgradedSendVoice: (
-            data: Uint8Array,
-            last: boolean,
-            messageName?: string,
-            params?: {
-                meta?: { [k: string]: string };
-            },
-        ) => void = (data, last, messageName, params) => {
-            checkLastMessageStatus(last);
-            return sendVoice(data, last, batchingMessageId, messageName, params);
-        };
-
-        return cb({
-            ...threeParamsMethods,
-            send: upgradedSend,
-            sendText: upgradedSendText,
-            sendSystemMessage: upgradedSendSystemMessage,
-            sendVoice: upgradedSendVoice,
-            messageId: batchingMessageId,
-        });
-    };
-
     return {
         once,
         send,
@@ -505,7 +223,7 @@ export const createClient = (clientParams: CreateClientDataType, logger?: Client
         waitForAnswerToUser,
         sendSystemMessage,
         on,
-        updateDefauls,
+        updateDefaults,
         destroy,
         batch,
         get currentMessageId() {
